[#7] Automatically save note on content change
