+++ package.json
@@ -15,7 +15,8 @@
   "dependencies": {
     "next": "13.5.6",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "use-debounce": "9.0.4"
   },
   "devDependencies": {
     "@playwright/test": "1.39.0",
